+++ src/Patient-Zero-main (Fixes)/Patient-Zero-main/src/Main.java
@@ -1,0 +1,13 @@
+import engine.GameEngine;
+import ui.TextRenderer;
+
+public class Main {
+    public static void main(String[] args) {
+        TextRenderer.clearScreen();
+           
+        GameEngine game = new GameEngine();
+        game.startGame();
+    }
+}
+
+