+++ src/Patient-Zero-main (Fixes)/Patient-Zero-main/src/effect/KnowledgeEffect.java
@@ -1,0 +1,17 @@
+package effect;
+
+import model.Player;
+
+//Effect that changes the players knowlegde
+public class KnowledgeEffect extends Effect {
+    // Removed private int amount
+
+    public KnowledgeEffect(int amount) {
+        super(amount);
+    }
+
+    @Override
+    public void apply(Player player) {
+        player.changeKnowledge(amount);
+    }
+}